Fix prints change properties to i32s

+++ src/community_cards.rs
@@ -1,5 +1,5 @@
 struct Card {
-    index: u8,
+    index: i32,
     name: String,
 
 }
